icons for light and darkmode ; Password visibility icons and ffunctionality ; Undefined error in My posts resolved

+++ src/pages/Post.jsx
@@ -11,9 +11,7 @@ export default function Post() {
     const [post, setPost] = useState(null);
     const { slug } = useParams();
     const navigate = useNavigate();
-
     const userData = useSelector((state) => state.auth.userData);
-
     const isAuthor = post && userData ? post.userId === userData.$id : false;
 
     useEffect(() => {
@@ -26,24 +24,29 @@ export default function Post() {
     }, [slug, navigate]);
 
     const deletePost = () => {
-        appwriteService.deletePost(post.$id).then((status) => {
-            if (status) {
-                appwriteService.deleteFile(post.featuredImage);
-                navigate("/");
-            }
-        });
+        const bool = confirm("Do you want to delete or clicked by mistake?")
+        if (bool){
+
+            appwriteService.deletePost(post.$id).then((status) => {
+                if (status) {
+                    appwriteService.deleteFile(post.featuredImage);
+                    navigate("/");
+                }
+            });
+        }else {
+            console.log("okay bye");
+        }
     };
 
     return post ? (
-        <div className="w-[70%] mx-auto my-4 p-16 rounded-3xl bg-slate-200 dark:text-gray-200 dark:bg-slate-900">
+        <div className="w-[70%] mx-auto my-4 p-16 rounded-3xl bg-slate-300 text-gray-900 dark:text-gray-200 dark:bg-slate-900">
             <Container>
-                <div className="w-full flex justify-center mb-4 relative border shadow-inner shadow-lime-200 rounded-xl p-8">
+                <div className="w-full flex justify-center mb-4 relative border border-gray-300 dark:border-gray-700 shadow-xl   shadow-gray-500 dark:shadow-lime-800 rounded-xl p-8 bg-white dark:bg-gray-800">
                     <img
                         src={appwriteService.getFilePreview(post.featuredImage)}
                         alt={post.title}
                         className="rounded-xl"
                     />
-
                     {isAuthor && (
                         <div className="absolute right-6 w-[30%] top-6 flex flex-wrap sm:flex-nowrap">
                             <Link to={`/edit-post/${post.$id}`} className="w-full sm:w-1/2 pr-1">
@@ -59,10 +62,10 @@ export default function Post() {
                         </div>
                     )}
                 </div>
-                <div className="w-full mb-6">
-                    <h1 className="text-center text-4xl mt-4 font-bold">{post.title}</h1>
+                <div className="w-full mt-20 mb-6">
+                    <h1 className="text-center text-4xl mt-4 font-bold text-gray-900 dark:text-gray-200">{post.title}</h1>
                 </div>
-                <div className="text-2xl browser-css">
+                <div className="text-2xl text-gray-900 dark:text-gray-200">
                     {parse(post.content)}
                 </div>
             </Container>
